+++ front/src/components/atoms/CheckBox.tsx
@@ -1,8 +1,7 @@
-import { forwardRef, Ref } from 'react'
+import { forwardRef, InputHTMLAttributes, Ref } from 'react'
 import styled from 'styled-components'
 import { colors, dimensions, FlexBox } from '../../styles'
 import Label from './Label'
-import { TCheckBox } from '../../types'
 
 export const CheckBoxWrapper = styled(FlexBox)`
   flex-direction: row;
@@ -40,7 +39,15 @@ export const CheckBoxWrapper = styled(FlexBox)`
     }
   }
 `
-
+type TCheckBox = InputHTMLAttributes<HTMLInputElement> & {
+  id: string
+  label: string
+  hiddenLabel?: boolean
+  defaultChecked?: boolean
+  required?: boolean
+  className?: string
+  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void
+}
 const CheckBox = forwardRef(
   (
     {
