+++ front/src/components/atoms/Input.tsx
@@ -1,7 +1,6 @@
-import React from 'react'
+import React, { InputHTMLAttributes } from 'react'
 import styled from 'styled-components'
 import { colors, dimensions, font } from '../../styles'
-import { TInput } from '../../types'
 
 const InputStyled = styled.input<TInput>`
   width: 100%;
@@ -19,7 +18,13 @@ const InputStyled = styled.input<TInput>`
     outline: 0 none;
   }
 `
-
+export type TInput = InputHTMLAttributes<HTMLInputElement> & {
+  error?: boolean | string
+  success?: boolean
+  warning?: boolean
+  type?: 'text' | 'password' | 'email'
+  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void
+}
 const Input = React.forwardRef<HTMLInputElement, TInput>(
   (
     {
