+++ front/src/components/atoms/Spinner.ts
@@ -1,6 +1,5 @@
 import styled, { keyframes } from 'styled-components'
 import { colors } from '../../styles'
-import { TSpinner } from '../../types'
 
 const spinnerAnimation = keyframes`
 from {
@@ -18,7 +17,9 @@ const sizes = {
   big: '133px',
 }
 
-
+type TSpinner = {
+  size?: 'xsmall' | 'small' | 'medium' | 'big'
+}
 const Spinner = styled.div.attrs<TSpinner>((props) => ({ size: props?.size })) <TSpinner>`
  ${({ size }) => `
     height: ${size ? sizes[size] : sizes.medium};
