+++ front/src/components/atoms/Textarea.tsx
@@ -1,7 +1,6 @@
-import React from 'react'
+import React, { TextareaHTMLAttributes } from 'react'
 import styled from 'styled-components'
 import { colors, dimensions, font } from '../../styles'
-import { TTextarea } from '../../types'
 
 const TextareaStyled = styled.textarea<TTextarea>`
   width: 100%;
@@ -18,7 +17,13 @@ const TextareaStyled = styled.textarea<TTextarea>`
     outline: 0 none;
   }
 `
-
+export type TTextarea = TextareaHTMLAttributes<HTMLTextAreaElement> & {
+  rows: number
+  cols?: number
+  error?: boolean | string
+  success?: boolean
+  warning?: boolean
+}
 const Textarea = React.forwardRef<HTMLTextAreaElement, TTextarea>(
   ({ cols = 1, ...rest }, ref) => (
     <TextareaStyled cols={cols} ref={ref} {...rest} />
