+++ front/src/components/atoms/UserImage.tsx
@@ -1,6 +1,5 @@
-import { FC } from 'react'
+import { FC, HTMLAttributes } from 'react'
 import styled from 'styled-components'
-import { TUserImage } from '../../types'
 
 const StyledUserImage = styled.img<TUserImage>`
   width: 97px;
@@ -8,7 +7,10 @@ const StyledUserImage = styled.img<TUserImage>`
   border-radius: 50%;
   object-fit: cover;
 `
-
+type TUserImage = HTMLAttributes<HTMLDivElement> & {
+  src: string
+  alt: string
+}
 const UserImage: FC<TUserImage> = ({ src, alt }) => (
   <StyledUserImage src={src} alt={alt} />
 )
