+++ front/src/components/atoms/ValidationMessage.tsx
@@ -1,6 +1,5 @@
 import styled from 'styled-components'
 import { colors, font } from '../../styles'
-import { TValidationMessage } from '../../types'
 
 type TValidationMessageStyled = {
   color: 'success' | 'error' | 'warning'
@@ -13,6 +12,10 @@ const ValidationMessageStyled = styled.p<TValidationMessageStyled>`
   font-size: ${font.xs};
   font-family: ${font.fontFamily};
 `
+export type TValidationMessage = {
+  text?: string
+  color?: 'success' | 'error' | 'warning'
+}
 
 const ValidationMessage = ({
   text = '',
