+++ front/src/components/molecules/AccessModalContent.tsx
@@ -2,7 +2,6 @@ import styled from 'styled-components'
 import { Button, Text, Title } from '../atoms'
 import { FlexBox, colors, dimensions, font } from '../../styles'
 import img from '../../assets/icons/lock-dynamic-color.svg'
-import { TAccessModal } from '../../types'
 
 const ImageStyled = styled.div`
   margin-top: 3.5rem;
@@ -28,7 +27,11 @@ const ButtonStyled = styled(Button)`
 const StyledText = styled(Text)`
   font-weight: ${font.medium};
 `
-
+type TAccessModal = {
+  handleAccessModal: () => void
+  handleLoginModal: () => void
+  handleRegisterModal: () => void
+}
 const AccessModalContent = ({
   handleAccessModal,
   handleLoginModal,
