+++ front/src/components/molecules/CardHome.tsx
@@ -3,7 +3,6 @@ import { useTranslation } from 'react-i18next'
 import { FlexBox, colors, dimensions, font } from '../../styles'
 import { Text } from '../atoms'
 import bgHomeCardCorner from '../../assets/icons/bg-home-card-corner.svg'
-import { TCardHome } from '../../types'
 
 const CardContainer = styled(FlexBox)`
   flex-direction: column;
@@ -55,7 +54,12 @@ const SubtitleStyled = styled(Text)`
   font-weight: ${font.regular};
   line-height: 1.3rem;
 `
-
+type TCardHome = {
+  cardTitle: string
+  cardSubtitle: string
+  indicator: string
+  icon: string
+}
 export const CardHome = ({
   cardTitle,
   cardSubtitle,
