+++ front/src/components/molecules/InputGroup.tsx
@@ -2,7 +2,8 @@ import { forwardRef, Ref } from 'react'
 import styled from 'styled-components'
 import { dimensions, FlexBox } from '../../styles'
 import { Icon, Input, Label, ValidationMessage } from '../atoms'
-import { TInputGroup } from '../../types'
+import { TInput } from '../atoms/Input'
+import { TValidationMessage } from '../atoms/ValidationMessage'
 
 const InputGroupStyled = styled.div`
   width: 100%;
@@ -34,7 +35,17 @@ const iconDataTestId = (rest: TObjectWithDataTestId) => {
     dataTestId.charAt(0).toUpperCase() + dataTestId.slice(1)
   return dataTestId ? `icon${upperCaseDataTestId}` : ''
 }
-
+type TInputGroup = {
+  id: string
+  name: string
+  label: string
+  validationType?: TValidationMessage['color']
+  validationMessage?: TValidationMessage['text']
+  icon?: string
+  className?: string
+  iconClick?: () => void
+  hiddenLabel?: boolean
+} & TInput
 const InputGroup = forwardRef(
   (
     {
