+++ front/src/components/molecules/SelectGroup.tsx
@@ -2,7 +2,8 @@ import { forwardRef } from 'react'
 import styled from 'styled-components'
 import { dimensions } from '../../styles'
 import { Label, ValidationMessage, Select } from '../atoms'
-import { TSelectGroup } from '../../types'
+import { TValidationMessage } from '../atoms/ValidationMessage'
+import { TSelect } from '../atoms/Select'
 
 const SelectGroupStyled = styled.div`
   ${ValidationMessage} {
@@ -11,6 +12,14 @@ const SelectGroupStyled = styled.div`
   }
   width: 100%;
 `
+export type TSelectGroup = {
+  id: string
+  name: string
+  label: string
+  hiddenLabel?: boolean
+  icon?: string
+  validationMessage?: TValidationMessage['text']
+} & TSelect
 
 const SelectGroup = forwardRef<HTMLSelectElement, TSelectGroup>(
   ({ label, id, name, hiddenLabel, icon, validationMessage, ...rest }, ref) => (
