+++ front/src/types/index.tsx
@@ -1,7 +1,6 @@
 import type {
   TFavorites,
   TGetTopics,
-  TGetTypes,
   TTopic,
   TUser,
   TFilters,
@@ -14,42 +13,13 @@ import type {
   TMappedTopics,
   TVoteCountResponse,
   TVoteMutationData,
-  TVoteCounter,
-  TTypesFilterWidget,
-  TError,
-  TSize,
-  TTooltip,
-  TButton,
-  TCheckBox,
-  TCounter,
-  TDropdown,
-  THamburgerMenu,
-  TIcon,
-  TInput,
-  TLabel,
-  TLink,
-  TRadio,
-  TSelect,
-  TSpinner,
-  TText,
-  TTextarea,
-  TTitle,
-  TUserImage,
-  TValidationMessage,
-  TAccessModal,
-  TCardHome,
-  TCardProfile,
-  TCreateAuthor,
-  TResourceFav,
-  TInputGroup,
   TResourceTitleLink,
-  TSelectGroup,
+  TGetTypes,
 } from './types'
 
 export type {
   TFavorites,
   TGetTopics,
-  TGetTypes,
   TTopic,
   TUser,
   TFilters,
@@ -62,34 +32,6 @@ export type {
   TMappedTopics,
   TVoteCountResponse,
   TVoteMutationData,
-  TVoteCounter,
-  TTypesFilterWidget,
-  TError,
-  TSize,
-  TTooltip,
-  TButton,
-  TCheckBox,
-  TCounter,
-  TDropdown,
-  THamburgerMenu,
-  TIcon,
-  TInput,
-  TLabel,
-  TLink,
-  TRadio,
-  TSelect,
-  TSpinner,
-  TText,
-  TTextarea,
-  TTitle,
-  TUserImage,
-  TValidationMessage,
-  TAccessModal,
-  TCardHome,
-  TCardProfile,
-  TCreateAuthor,
-  TResourceFav,
-  TInputGroup,
   TResourceTitleLink,
-  TSelectGroup,
+  TGetTypes,
 }
